Replace useBoundingClientRect with useBoundingRect
The new version of the hook added the logic of measuring element after user scrolled. With new implementation, vertical position values(top, bottom, y) is always relative to the top of the page, instead of window top. This is done to simplify the common use case of manipulating page element with consideration of its scroll position.