Add iPad model, texture swapping, and scene.
Opted to go with a solution where the screens are different image textures rather than a plane that sits on top of the screen.
This way, the lighting reflections will apply to the screen itself.
Created a custom material with similar properties as the glass underneath, so that the reflections will be the same.

Given Figma screens were also not designed to scale unfortunately.
I had to go back and adjust the model to fit the sizing and create custom UV screen plane that matched the given mockup.

Resources:
- Drei: https://github.com/pmndrs/drei
- R3F: https://docs.pmnd.rs/react-three-fiber/
- Texture Flip Y: https://threejs.org/docs/#api/en/textures/Texture.flipY
- Linear interpolation formula: https://graphics.fandom.com/wiki/Linear_interpolation